Fetching questionnaire if needed
