+++ web/static/js/actions/questionnaire.js
@@ -1,4 +1,4 @@
-import { createQuestionnaire, updateQuestionnaire } from '../api'
+import * as api from '../api'
 import { buildNewStep } from '../reducers/questionnaire'
 
 export const NEW = 'QUESTIONNAIRE_NEW'
@@ -14,6 +14,42 @@ export const CHANGE_STEP_STORE = 'QUESTIONNAIRE_CHANGE_STEP_STORE'
 export const ADD_CHOICE = 'QUESTIONNAIRE_EDITOR_ADD_CHOICE'
 export const DELETE_CHOICE = 'QUESTIONNAIRE_EDITOR_DELETE_CHOICE'
 export const CHANGE_CHOICE = 'QUESTIONNAIRE_EDITOR_CHANGE_CHOICE'
+
+export const fetchQuestionnaire = (projectId, questionnaireId) => (dispatch, getState) => {
+  dispatch(fetch(projectId, questionnaireId))
+  return api.fetchQuestionnaire(projectId, questionnaireId)
+    .then(response => {
+      dispatch(receive(response.entities.questionnaires[response.result]))
+    })
+    .then(() => {
+      return getState().questionnaire.data
+    })
+}
+
+export const fetch = (projectId, questionnaireId) => ({
+  type: FETCH,
+  projectId,
+  questionnaireId
+})
+
+export const fetchQuestionnaireIfNeeded = (projectId, questionnaireId) => {
+  return (dispatch, getState) => {
+    if (shouldFetch(getState().questionnaire, projectId, questionnaireId)) {
+      return dispatch(fetchQuestionnaire(projectId, questionnaireId))
+    } else {
+      return Promise.resolve(getState().questionnaire.data)
+    }
+  }
+}
+
+export const receive = (questionnaire) => ({
+  type: RECEIVE,
+  questionnaire
+})
+
+export const shouldFetch = (state, projectId, questionnaireId) => {
+  return !state.fetching || !(state.filter && (state.filter.projectId === projectId && state.filter.questionnaireId === questionnaireId))
+}
 
 export const addChoice = (stepId) => ({
   type: ADD_CHOICE,
@@ -60,17 +96,6 @@ export const newQuestionnaire = (projectId) => ({
   projectId
 })
 
-export const receive = (questionnaire) => ({
-  type: RECEIVE,
-  questionnaire
-})
-
-export const fetch = (projectId, questionnaireId) => ({
-  type: FETCH,
-  projectId,
-  questionnaireId
-})
-
 export const changeName = (newName) => ({
   type: CHANGE_NAME,
   newName
@@ -93,9 +118,9 @@ export const save = () => {
     const questionnaire = getState().questionnaire
 
     if (questionnaire.id == null) {
-      createQuestionnaire(questionnaire.projectId, questionnaire)
+      api.createQuestionnaire(questionnaire.projectId, questionnaire)
     } else {
-      updateQuestionnaire(questionnaire.projectId, questionnaire)
+      api.updateQuestionnaire(questionnaire.projectId, questionnaire)
     }
   }
 }
